pyproject.toml
[tool.poetry]
name = "hotel-reservation"
version = "0.1.0"
description = ""
authors = ["Shrine Sabu <shrine02sabu@gmail.com>"]
readme = "README.md"
packages = [{include = "hotel_reservation"}]

[tool.poetry.dependencies]
python = "3.10"
django = "^5.0.7"
djangorestframework = "^3.15.2"
celery = "^5.4.0"
redis = "^5.0.8"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
